+++ src/components/App.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import { Button } from '@material-ui/core';
 
 const getMedia = async () => {
   const constraints = { audio: true, video: true };
@@ -13,7 +14,13 @@ const getMedia = async () => {
 getMedia();
 
 const App = () => {
-  return <div>App</div>;
+  return (
+    <>
+      <Button color="primary" variant="outlined">
+        Hello, World!
+      </Button>
+    </>
+  );
 };
 
 export default App;
